+++ tests/functionality/test_fb_oauth.py
