+++ tests/functionality/test_get_session_by_id.py
