+++ tests/functionality/test_get_speaker_by_id.py
