+++ tests/functionality/test_session_type.py
