+++ tests/views/admin/test_my_session.py
