+++ tests/views/admin/test_speakers.py
