+++ tests/views/guest/test_guest_event_page.py
